"Refactor slice utilities into a dedicated file"
Moved slice-related utilities from `godash.go` to a new `slices.go` file for better modularity and readability. Updated file structure by renaming `godash_test.go` to `slices_test.go` to reflect changes appropriately. This refactoring isolates slice operations while maintaining functionality.
